feat (src/tools/rag/): add langchain tools retrieve query
