+++ src/tools/rag/__init__.py
@@ -1,0 +1,4 @@
+#
+#  Copyright (c) 2025
+#  Minh NGUYEN <vnguyen9@lakeheadu.ca>
+#
